splitting client back out
